requirements.txt
boto3 ==1.15.18
mxnet==1.7.0.post2
torch==1.12
urllib3==1.25.4
onnxruntime==1.8.0
transformers==4.8.2
sentencepiece==0.1.99
gluonnlp==0.8.0
pandas==1.2.1
numpy==1.17
haversine

fastapi==0.101.0
uvicorn==0.22.0
jinja2==3.1.2
pydantic==2.1.1
faiss-cpu==1.7.4